Se  crean las nuevas interfaces de la fase 2 y se muetra el primer formulario de validacion

+++ FactorPolar.Domain/Entities/RubricaEvaluacion.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string? Descripcion { get; set; }
-        public int Factor { get; set; }
+        public decimal Factor { get; set; }
         public DateTime CreateDate { get; set; }
         public DateTime UpdateDate { get; set; }
